Crafting System Part 3.1 - Item Selection from Bag and Debug Menu
Added the logic for selecting a slot, being taken to the bag, selecting an item and quantity, and storing that info for use in the crafting system. These items are displayed on the crafting table but I've also added a debug menu to list the items and quantities in each slot. There's a few commits before this that are also part of part 3 that I'm not able to include in this squash.