Crafting System Part 3.3 - Get pack up working
When trying to close crafting menu while items are on the table, it will now ask if you want to pack up with a yes or no selection window. If you say yes, the crafting menu will close and the items will be put back into the bag. If you say no, you'll return to the crafting menu.