Crafting System Part 4 - Action Menu
Selecting a slot with an item will now bring up an action menu where you can choose the following options and their intended features:
Swap item - open the bag and allow you to choose an item which will replace the item you had in that slot while the old item is put back in the bag.
Adjust quantity - Then prompt then to ask how many they'd like to have placed on the table. This should still be limited to the amount they actually have in their bag of that item
Swap Slot - move the selected item and quantity to a different slot on the table. Maybe the cursor should change color while they are selecting a new slot.
Remove item - place the selected item back into the bag
Cancel - close the submenu